Handle AI errors, refactor
