+++ test/sampleDirectories/kitchenSink/script.ts
@@ -1,2 +1,4 @@
 const testing = 500;
 //Add 20 to testing and print it out
+const result = testing + 20;
+console.log(result); // should output 520
